server/AI/requirements.txt
flask
tensorflow==2.20.0
numpy
opencv-python-headless
matplotlib
pandas
tabulate
Pillow
cv2
io
os